move to new flash_mla module build
