+++ csrc/xpu/attn/attn_interface.cpp
@@ -2,46 +2,10 @@
 #include "attn_interface.h"
 
 #ifdef VLLM_FA2_ATTN_ENABLED
-#ifdef VLLM_XPU_ENABLE_XE2
-  #include "csrc/xpu/attn/xe_2/fmha_xe2.h"
-  #include "csrc/xpu/attn/xe_2/paged_decode_xe2.h"
-#endif
-#endif
-
-#ifdef VLLM_SPARSE_MLA_ENABLED
-
-namespace vllm::xpu::attn::mla {
-
-std::vector<at::Tensor> flash_mla_sparse_attn_prefill_impl(
-    const at::Tensor& q,
-    const at::Tensor& kv,
-    const at::Tensor& indices,
-    float sm_scale,
-    int d_v,
-    const std::optional<at::Tensor>& attn_sink,
-    const std::optional<at::Tensor>& topk_length,
-    const std::optional<at::Tensor>& output,
-    bool return_softmax_lse);
-
-std::tuple<at::Tensor, std::optional<at::Tensor>>
-flash_mla_sparse_attn_decode_impl(
-    const at::Tensor& q,
-    const at::Tensor& kv,
-    const at::Tensor& indices,
-    const std::optional<at::Tensor>& topk_length,
-    const std::optional<at::Tensor>& attn_sink,
-    const std::optional<at::Tensor>& out,
-    std::optional<at::Tensor>& tile_scheduler_metadata,
-    std::optional<at::Tensor>& num_splits,
-    const std::optional<at::Tensor>& extra_kv,
-    const std::optional<at::Tensor>& extra_indices,
-    const std::optional<at::Tensor>& extra_topk_length,
-    int d_v,
-    float sm_scale,
-    bool return_softmax_lse);
-
-}  // namespace vllm::xpu::attn::mla
-
+  #ifdef VLLM_XPU_ENABLE_XE2
+    #include "csrc/xpu/attn/xe_2/fmha_xe2.h"
+    #include "csrc/xpu/attn/xe_2/paged_decode_xe2.h"
+  #endif
 #endif
 
 #ifdef VLLM_FA2_ATTN_ENABLED
@@ -71,7 +35,7 @@ void cutlass_chunk_prefill_interface(
     std::optional<at::Tensor>& softmax_lse,
     std::optional<const at::Tensor>& is_prefill) {
   if (vllm::xpu::is_xe2_arch() || vllm::xpu::is_xe3_arch()) {
-#ifdef VLLM_XPU_ENABLE_XE2
+  #ifdef VLLM_XPU_ENABLE_XE2
     // Use XE2 cutlass kernel (also used as WA for XE3/XE3P)
     cutlass_chunk_prefill_xe2(
         queue,
@@ -97,9 +61,9 @@ void cutlass_chunk_prefill_interface(
         is_sink,
         softmax_lse,
         is_prefill);
-#else
+  #else
     TORCH_CHECK(false, "XE2 cutlass kernel is not enabled in this build.");
-#endif
+  #endif
   } else {
     TORCH_CHECK(false, "Only XE2/XE3 cutlass kernel is supported currently.");
   }
@@ -136,7 +100,7 @@ void cutlass_paged_decode_interface(
     std::optional<at::Tensor>& splits_per_seq,
     std::optional<at::Tensor>& work_list) {
   if (vllm::xpu::is_xe2_arch() || vllm::xpu::is_xe3_arch()) {
-#ifdef VLLM_XPU_ENABLE_XE2
+  #ifdef VLLM_XPU_ENABLE_XE2
     // Use XE2 cutlass kernel (also used as WA for XE3/XE3P)
     cutlass_paged_decode_xe2(
         queue,
@@ -167,70 +131,12 @@ void cutlass_paged_decode_interface(
         is_prefill,
         splits_per_seq,
         work_list);
-#else
+  #else
     TORCH_CHECK(false, "XE2 cutlass kernel is not enabled in this build.");
-#endif
+  #endif
   } else {
     TORCH_CHECK(false, "Only XE2/XE3 cutlass kernel is supported currently.");
   }
-}
-
-#endif
-
-#ifdef VLLM_SPARSE_MLA_ENABLED
-std::vector<at::Tensor> flash_mla_sparse_attn_prefill_interface(
-    const at::Tensor& q,
-    const at::Tensor& kv,
-    const at::Tensor& indices,
-    float sm_scale,
-    int d_v,
-    const std::optional<at::Tensor>& attn_sink,
-    const std::optional<at::Tensor>& topk_length,
-    const std::optional<at::Tensor>& output,
-    bool return_softmax_lse) {
-  return vllm::xpu::attn::mla::flash_mla_sparse_attn_prefill_impl(
-      q,
-      kv,
-      indices,
-      sm_scale,
-      d_v,
-      attn_sink,
-      topk_length,
-      output,
-      return_softmax_lse);
-}
-
-std::tuple<at::Tensor, std::optional<at::Tensor>>
-flash_mla_sparse_attn_decode_interface(
-    const at::Tensor& q,
-    const at::Tensor& kv,
-    const at::Tensor& indices,
-    const std::optional<at::Tensor>& topk_length,
-    const std::optional<at::Tensor>& attn_sink,
-    const std::optional<at::Tensor>& out,
-    std::optional<at::Tensor>& tile_scheduler_metadata,
-    std::optional<at::Tensor>& num_splits,
-    const std::optional<at::Tensor>& extra_kv,
-    const std::optional<at::Tensor>& extra_indices,
-    const std::optional<at::Tensor>& extra_topk_length,
-    int d_v,
-    float sm_scale,
-    bool return_softmax_lse) {
-  return vllm::xpu::attn::mla::flash_mla_sparse_attn_decode_impl(
-      q,
-      kv,
-      indices,
-      topk_length,
-      attn_sink,
-      out,
-      tile_scheduler_metadata,
-      num_splits,
-      extra_kv,
-      extra_indices,
-      extra_topk_length,
-      d_v,
-      sm_scale,
-      return_softmax_lse);
 }
 
 #endif
