+++ csrc/xpu/attn/attn_interface.h
@@ -62,31 +62,3 @@ void cutlass_paged_decode_interface(
     std::optional<const at::Tensor>& is_prefill,
     std::optional<at::Tensor>& splits_per_seq,
     std::optional<at::Tensor>& work_list);
-
-std::vector<at::Tensor> flash_mla_sparse_attn_prefill_interface(
-	const at::Tensor& q,
-	const at::Tensor& kv,
-	const at::Tensor& indices,
-	float sm_scale,
-	int d_v,
-	const std::optional<at::Tensor>& attn_sink,
-	const std::optional<at::Tensor>& topk_length,
-	const std::optional<at::Tensor>& output,
-	bool return_softmax_lse);
-
-std::tuple<at::Tensor, std::optional<at::Tensor>>
-flash_mla_sparse_attn_decode_interface(
-	const at::Tensor& q,
-	const at::Tensor& kv,
-	const at::Tensor& indices,
-	const std::optional<at::Tensor>& topk_length,
-	const std::optional<at::Tensor>& attn_sink,
-	const std::optional<at::Tensor>& out,
-	std::optional<at::Tensor>& tile_scheduler_metadata,
-	std::optional<at::Tensor>& num_splits,
-	const std::optional<at::Tensor>& extra_kv,
-	const std::optional<at::Tensor>& extra_indices,
-	const std::optional<at::Tensor>& extra_topk_length,
-	int d_v,
-	float sm_scale,
-	bool return_softmax_lse);
