--- csrc/xpu/attn/kernel_configs/sparse_mla_default.conf
@@ -1,14 +1,0 @@
-# =============================================================================
-# Default sparse MLA kernel configuration
-# =============================================================================
-# Format:
-#   prefill,headsize[,topklen[,attn_sink]]
-#   decode_fp8,headsize[,topklen[,attn_sink]]
-#
-# If bool fields are omitted, all combinations are generated for that entry.
-
-# Prefill kernels (DeepSeek V4): head_dim_qk 512, with topk_length specialization.
-prefill,512,true,true
-
-# Decode FP8 kernels (DeepSeek V4): head_dim_qk 512, with topk_length specialization.
-decode_fp8,512,true,true
