+++ csrc/xpu/flash_mla/kernel_configs/sparse_mla_default.conf
@@ -1,5 +1,6 @@
 # =============================================================================
-# Full sparse MLA kernel configuration
+# Default sparse MLA kernel configuration
 # =============================================================================
-# Build all sparse MLA kernel instantiations.
+# Keep default aligned with full coverage so users do not need to choose
+# between "all" and "default".
 all
