+++ csrc/xpu/flash_mla/mla_interface.h
@@ -1,0 +1,35 @@
+#pragma once
+
+#include <optional>
+#include <tuple>
+#include <vector>
+
+#include <ATen/ATen.h>
+
+std::vector<at::Tensor> flash_mla_sparse_attn_prefill_interface(
+    const at::Tensor& q,
+    const at::Tensor& kv,
+    const at::Tensor& indices,
+    float sm_scale,
+    int d_v,
+    const std::optional<at::Tensor>& attn_sink,
+    const std::optional<at::Tensor>& topk_length,
+    const std::optional<at::Tensor>& output,
+    bool return_softmax_lse);
+
+std::tuple<at::Tensor, std::optional<at::Tensor>>
+flash_mla_sparse_attn_decode_interface(
+    const at::Tensor& q,
+    const at::Tensor& kv,
+    const at::Tensor& indices,
+    const std::optional<at::Tensor>& topk_length,
+    const std::optional<at::Tensor>& attn_sink,
+    const std::optional<at::Tensor>& out,
+    std::optional<at::Tensor>& tile_scheduler_metadata,
+    std::optional<at::Tensor>& num_splits,
+    const std::optional<at::Tensor>& extra_kv,
+    const std::optional<at::Tensor>& extra_indices,
+    const std::optional<at::Tensor>& extra_topk_length,
+    int d_v,
+    float sm_scale,
+    bool return_softmax_lse);
